@@ -10,9 +10,9 @@
     "enzyme-adapter-react-16": "1.7.0",
     "flow-bin": "^0.63.1",
     "flow-result-checker": "^0.3.0",
-    "jest-expo": "^38.0.0",
+    "jest-expo": "^39.0.0",
     "jest-fetch-mock": "1.7.4",
-    "react-dom": "16.11.0",
+    "react-dom": "16.13.1",
     "react-test-renderer": "16.0.0-alpha.12",
     "redux-mock-store": "1.5.3",
     "remote-redux-devtools": "^0.5.12"
@@ -36,19 +36,22 @@
     ]
   },
   "dependencies": {
-    "@react-native-community/datetimepicker": "2.4.0",
+    "@react-native-async-storage/async-storage": "1.13.2",
+    "@react-native-community/datetimepicker": "3.0.0",
     "@react-native-community/masked-view": "0.1.10",
-    "@react-native-community/netinfo": "5.9.2",
+    "@react-native-community/netinfo": "5.9.6",
+    "@react-native-picker/picker": "1.9.4",
     "android-versions": "^1.2.1",
     "autobind-decorator": "^2.1.0",
     "babel-preset-es2015": "^6.24.1",
     "crypto-js": "3.1.9-1",
     "dotenv": "8.2.0",
-    "expo": "^38.0.0",
+    "expo": "^39.0.0",
     "expo-analytics": "^1.0.7",
+    "expo-clipboard": "1.0.1",
     "expo-linking": "~1.0.1",
-    "expo-permissions": "~9.0.1",
-    "expo-store-review": "~2.1.2",
+    "expo-permissions": "~9.3.0",
+    "expo-store-review": "~2.2.0",
     "firebase": "7.9.0",
     "formik": "1.4.1",
     "hoist-non-react-statics": "^2.3.1",
@@ -61,25 +64,25 @@
     "node-emoji": "^1.8.1",
     "npm": "^5.4.2",
     "prop-types": "^15.6.0",
-    "react": "16.11.0",
+    "react": "16.13.1",
     "react-moment": "^0.6.5",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.3.tar.gz",
     "react-native-check-box": "^2.0.2",
     "react-native-collapsible": "^0.9.0",
     "react-native-dropdownalert": "^3.4.0",
     "react-native-elements": "^0.19.1",
     "react-native-extended-stylesheet": "^0.6.0",
-    "react-native-gesture-handler": "~1.6.0",
-    "react-native-keyboard-aware-scroll-view": "^0.3.0",
+    "react-native-gesture-handler": "~1.7.0",
+    "react-native-keyboard-aware-scroll-view": "^0.9.2",
     "react-native-modal": "^4.1.1",
     "react-native-modalbox": "^1.4.2",
     "react-native-platform-touchable": "^1.1.1",
-    "react-native-reanimated": "~1.9.0",
-    "react-native-safe-area-context": "~3.0.7",
-    "react-native-screens": "~2.9.0",
+    "react-native-reanimated": "~1.13.0",
+    "react-native-safe-area-context": "3.1.4",
+    "react-native-screens": "~2.10.1",
     "react-native-searchbar": "^1.15.0",
     "react-native-svg": "12.1.0",
-    "react-native-webview": "9.4.0",
+    "react-native-webview": "10.7.0",
     "react-navigation": "^3.11.1",
     "react-navigation-backhandler": "1.2.0",
     "react-redux": "^5.0.6",
@@ -91,7 +94,7 @@
     "redux-saga": "^0.15.3",
     "redux-thunk": "^2.2.0",
     "sentry-expo": "~2.0.0",
-    "tcomb-form-native": "^0.6.10",
+    "tcomb-form-native": "0.6.20",
     "yup": "0.26.6"
   },
   "rnpm": {
