@@ -1,5 +1,6 @@
 import jwtDecode from "../../node_modules/jwt-decode";
-import { AsyncStorage } from "react-native";
+import AsyncStorage from '@react-native-async-storage/async-storage';
+
 import * as Font from 'expo-font'
 import { takeEvery, takeLatest, select, call, put } from "redux-saga/effects";
 import Environment from "../config/environment";
@@ -427,25 +428,6 @@ function* updateUser() {
           Authorization: "Bearer " + token
         },
         body: JSON.stringify(user)
-
-        // body: JSON.stringify({
-        //   gamertag: user.gamertag,
-        //   platform: user.platform,
-        //   play_style: user.play_style,
-        //   play_schedule: user.play_schedule,
-        //   light_level: user.light_level,
-        //   age: user.age,
-        //   no_emails: user.no_emails,
-        //   no_push_notifications: user.no_push_notifications,
-        //   push_new_group_game: user.push_new_group_game,
-        //   push_new_friend_game: user.push_new_friend_game,
-        //   push_player_joined_left: user.push_player_joined_left,
-        //   push_game_time_changed: user.push_game_time_changed,
-        //   push_username_mention: user.push_username_mention,
-        //   push_karma_received: user.push_karma_received,
-        //   push_private_message_received: user.push_private_message_received,
-        //   push_game_reminder: user.push_game_reminder
-        // })
       }
     );
     const result = yield response.json();
@@ -469,9 +451,6 @@ function* createGamingSession() {
     let gamingSession = yield select(
       state => state.gamingSessions.gamingSession
     );
-    let gamingSessionVisibility = yield select(
-      state => state.gamingSessions.gamingSessionVisibility
-    );
     let platform = yield select(state => state.search.platform);
 
     const response = yield fetch(
@@ -485,26 +464,8 @@ function* createGamingSession() {
           "Content-Type": "application/json",
           Authorization: "Bearer " + token
         },
-        body: JSON.stringify({
-          game_id: gamingSession.game_id,
-          platform: platform,
-          description: gamingSession.description,
-          activity: gamingSession.activity,
-          start_time: gamingSession.start_time,
-          group_name: gamingSession.group ? gamingSession.group : null,
-          make_auto_public: gamingSession.make_auto_public,
-          beginners_welcome: gamingSession.beginners_welcome,
-          sherpa_requested: gamingSession.sherpa_requested,
-          headset_required: gamingSession.mic_required,
-          party_size: gamingSession.party_size,
-          platform: gamingSession.platform,
-          created_from: "mobile-app",
-          public_visible: gamingSessionVisibility.publicVisible,
-          alliance_visible: gamingSessionVisibility.allianceVisible,
-          group_visible: gamingSessionVisibility.groupVisible,
-          friends_visible: gamingSessionVisibility.friendsVisible,
-          private_visible: gamingSessionVisibility.privateVisible
-        })
+        body: JSON.stringify(gamingSession)
+
       }
     );
     console.log(response);
@@ -530,9 +491,7 @@ function* editGamingSession() {
     let gamingSessionId = yield select(
       state => state.gamingSessions.gamingSessionId
     );
-    let gamingSessionVisibility = yield select(
-      state => state.gamingSessions.gamingSessionVisibility
-    );
+
     let platform = yield select(state => state.search.platform);
     console.log(
       Environment["API_BASE_URL"] +
@@ -540,6 +499,7 @@ function* editGamingSession() {
       "gaming_sessions/" +
       gamingSessionId
     );
+
     const response = yield fetch(
       Environment["API_BASE_URL"] +
       Environment["API_VERSION"] +
@@ -552,27 +512,8 @@ function* editGamingSession() {
           "Content-Type": "application/json",
           Authorization: "Bearer " + token
         },
-        body: JSON.stringify({
-          game_id: gamingSession.game_id,
-          name: gamingSession.description,
-          category: gamingSession.activity,
-          platform: platform,
-          start_time: gamingSession.start_time,
-          start_date: gamingSession.start_time,
-          group_name: gamingSession.group,
-          // friends_only: gamingSession.friends_only,
-          // group_only: gamingSession.group_only,
-          make_auto_public: gamingSession.make_auto_public,
-          beginners_welcome: gamingSession.beginners_welcome,
-          sherpa_requested: gamingSession.sherpa_requested,
-          mic_required: gamingSession.mic_required,
-          party_size: gamingSession.party_size,
-          platform: gamingSession.platform,
-          public_visible: gamingSessionVisibility.publicVisible,
-          group_visible: gamingSessionVisibility.groupVisible,
-          friends_visible: gamingSessionVisibility.friendsVisible,
-          private_visible: gamingSessionVisibility.privateVisible
-        })
+        body: JSON.stringify(gamingSession)
+
       }
     );
     const result = yield response.json();
@@ -931,7 +872,7 @@ function* fetchGroup() {
         Environment["API_BASE_URL"] +
         Environment["API_VERSION"] +
         "groups/" +
-        user.groups[0]["id"];
+        user.groups[0]["id"] + "/show_public";
       console.log("FETCHING GROUP (no default group): ", endpoint);
 
       yield call(fetchData, endpoint, 1, FETCH_GROUP_RESULT, FETCH_GROUP_ERROR);
@@ -943,10 +884,10 @@ function* fetchGroup() {
         Environment["API_BASE_URL"] +
         Environment["API_VERSION"] +
         "groups/" +
-        selectedGroupId;
+        selectedGroupId + "/show_public";
       console.log("FETCHING GROUP: ", endpoint);
 
-      yield call(fetchData, endpoint, 1, FETCH_GROUP_RESULT, FETCH_GROUP_ERROR);
+      yield call(fetchData, endpoint, null, FETCH_GROUP_RESULT, FETCH_GROUP_ERROR);
       AsyncStorage.setItem("default_group_id", selectedGroupId.toString());
     }
   } catch (e) {
